낙관적 락 진행

+++ src/main/java/org/hanghae99/company/product/entity/Product.java
@@ -27,7 +27,7 @@ public class Product extends TimeStamped {
     private Double score;
 
     @Version
-    private Integer version;
+    private Long version;
 
     @OneToMany(mappedBy = "product", cascade = CascadeType.PERSIST)
     private List<Review> reviews  = new ArrayList<>();
